+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Support for Nyaa.si tracker.
 - Support for Limetorrents tracker.
 - Set download path for downloading torrent files.
+- Automatically get working proxy ThePirateBay, KickAssTorrents, LimeTorrents.
 
 ## [1.0.3] - 2018-03-04
 ### Fixed